+++ get_tube_data.pl
@@ -5,7 +5,7 @@ use Math::Trig;
 use Math::Complex;
 use List::Util qw[min max];
 
-my $length_bound = 2.0;
+my $length_bound = 1.5;
 my $ortho_bound = 8.0;
 my $type = 'closed';
 my $mfld_count  = 11031;
@@ -25,6 +25,7 @@ if (@ARGV > 0) {
 sub get_geods {
     my ($t, $n, $l) = @_;
     return  "echo 'read $t $n
+                   print name
                    print geodesics $l' | snap |";
 }
 
@@ -32,12 +33,15 @@ sub get_ortho {
     my ($t, $n, $l, $d, @gs)  = @_;
     return "echo 'read $t $n
                   print geodesics $l
-                  print ortholines $d @gs' | snap |";
+                  ortholines $d @gs' | snap |";
 }
+
+print "\"name\",\"geod\",\"length\",\"ortho\",\"move\"\n";
 
 MAIN: 
 for my $n (1 .. $mfld_count) {
-    print STDERR "$type $n\n";
+    my $name = "$type $n";
+    print STDERR "$name\n";
     my %geods;
     # Simple, but hang suseptable way of proecessing system command
     open(PS, get_geods($type, $n, $length_bound)) or die "Failed on : $!";
@@ -47,6 +51,8 @@ for my $n (1 .. $mfld_count) {
             my $real = $2;
             my $imag = $3;
             $geods{$m} = $real + $imag * i;
+        } elsif (/.*name : (.*)/) {
+            $name = $1;
         } elsif (/.*Dirichlet.*/) {
             print STDERR "Dirichlet domain failed for $type $n\n";
             next MAIN;
@@ -57,15 +63,15 @@ for my $n (1 .. $mfld_count) {
     if (keys %geods < 1) { next MAIN; }
 
     # Different/safer way of processing system command
-    my $pid;
     my %orthos;
     INNER:
     foreach my $g (keys %geods) {
         my $ortho_len = max(Re($geods{$g}),2); 
         while ((! exists $orthos{$g}) and ($ortho_len < $ortho_bound)) { 
+            my $pid;
             eval {
                 local $SIG{ALRM} = sub { die "alarm\n" }; # NB: \n required
-                alarm 180;
+                alarm 240;
                 $pid = open(PS, get_ortho($type, $n, $length_bound, $ortho_len, $g)) or die "Failed on : $!";
                 while (<PS>) {
                     if (/([0-9\.]+)([0-9\.\+\-]+)\*i  (\d+):.+  (\d+):.*/) {
@@ -73,7 +79,7 @@ for my $n (1 .. $mfld_count) {
                         my $imag = $2;
                         my $left = $3;
                         my $right = $4; 
-                        if (($left == $right) and (! exists $orthos{$left})) {
+                        if (($left == $right) and ((! exists $orthos{$left}) or ($real < Re($orthos{$left})))) {
                             if ($real != 0.0) {
                                 $orthos{$left} = $real + $imag * i; 
                             } else {
@@ -82,12 +88,17 @@ for my $n (1 .. $mfld_count) {
                         } 
                     }
                 }
-                close PS;
                 alarm 0;
+                close PS;
             };
+            alarm 0; # race contition 
             if ($@) {
-                print STDERR "Timeout for $type $n\n";
-                kill 9, $pid+2; # HACK!!! echo and snap in the command get their own pids.
+                if ($@ eq "alarm\n") {
+                    print STDERR "Timeout for $type $n\n";
+                } else {
+                    print STDERR "Failed for $type $n\n";
+                }            
+                if (defined $pid) { kill 9, $pid+2; } # HACK!!! echo and snap in the command get their own pids.
                 next INNER;
             }
             $ortho_len += 0.5;
@@ -102,7 +113,7 @@ for my $n (1 .. $mfld_count) {
             $d = min($d, acosh(cosh(Re($k*$l))*cosh(Re($r)/2)**2 - cos(Im($k*$l))*sinh(Re($r)/2)**2));
             $k += 1;
         }
-        my $out = "$type $n, geod $g, length $l, ortho $r, move $d\n";
+        my $out = "\"$name\",$g,$l,$r,$d\n";
         $out =~ s/i/*1j/g;
         print "$out";
     }
